testNicolas/frontend/requirements.txt
altgraph==0.17.5
certifi==2026.2.25
charset-normalizer==3.4.7
idna==3.11
packaging==26.1
pefile==2024.8.26
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.4
PySide6==6.11.0
PySide6_Addons==6.11.0
PySide6_Essentials==6.11.0
python-dotenv==1.2.2
pywin32-ctypes==0.2.3
requests==2.33.1
shiboken6==6.11.0
urllib3==2.6.3
